pyproject.toml
[project]
name = "captcha-recognition"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "captcha>=0.7.1",
    "datasets>=4.5.0",
    "huggingface-hub>=1.3.7",
    "kagglehub>=0.4.2",
    "python-dotenv>=1.2.1",
    "python-levenshtein>=0.27.3",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
    "tensorflow[and-cuda]>=2.20.0",
    "torch>=2.9.1",
    "torchao>=0.16.0",
    "torchinfo>=1.8.0",
    "torchvision>=0.24.1",
    "transformers>=5.1.0",
    "wandb>=0.25.0",
]

[[tool.uv.index]] 
name = "pytorch-cu130" 
url = "https://download.pytorch.org/whl/cu130" 
explicit = true 

[tool.uv.sources] 
torch = { index = "pytorch-cu130" } 
torchvision = { index = "pytorch-cu130" }

[dependency-groups]
dev = [
    "jupyter>=1.1.1",
    "matplotlib>=3.10.8",
]
